+++ logo/logo.py
@@ -1,46 +1,40 @@
 import time
+from logo.color import *
 
-class bcolors:
-    MAGENTA = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKGREEN = '\033[92m'
-    ORANGE = '\033[93m'
-    RED = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
+def logo():
+    print ("\n")
+    print (bcolors.OKBLUE + "                                            @@@@@@@@   @@@@@@@@@         ")
+    print ("                             @@@@@@@      @@@@@@@@@@@ @@@@@@@@@@@@             ")
+    print ("                           @@@@@@@@@@@@  @@@@@@ @@@@@ @@@@@@@@@@@@@  @@@@      ")
+    print ("  @@@@          @@@@@@@@@@ @@@@@@@@@@@@@ @@@@@@ @@@@  @@@@@%   @@@ @@@@@@@@@   ")
+    print ("@@@@@@@@@    @@@@@@@@@@@@@ #@@@@@@@@@@@@ @@@@@@      @@@@@@@     @@@@@@@@@@@@@ ")
+    print ("@@@@@@@@@@   @@@@@@@@@@@@  @@@@@  @@@@@ #&&&&&&&&    @@@@@@@@@@ @@@@@@ @@@@@@@@")
+    print (" @@@@@@@@@@   @@@@@@@@@@@@  @@@@@  @@@@@ #&&&&&&&&   @@@@@@@@@@ @@@@@   @@@@@@@")
+    print ("  @@@@@@@@@@&  @@@@@@@@@    @@@@&  &&&&&  &&&&&&&&% &&&&&&&&@  &@@@@    @@@@@@@")
+    print ("   @@@@  %@&&&  @@@&&&       &&&&&  &&&&    &&&&&&& &&&&      &&&&@     (@@@@# ")
+    print ("    @@@&   &&&&  &&&&&.      &&&&&  &&&&      %%%% %&&&      &&&&              ")
+    print ("     &&&&   &&&%  &&%%%%%%%%  %%%%  %%%% /%%   %%% %%%%%%%  %&&& &&&&&&        ")
+    print ("      &&&&   %%%#  %%%%%%%%#  %%%%  %%%% %%%%%%%%% %%%%%%% %%%%&&&&&&          ")
+    print ("       %%%%   %%%   %######   ##### #### ########  ##%%%%  %%%%%%%&            ")
+    print ("        %%%%   ###  #####      #### #### (######             %%%%              ")
+    print ("         %###  ####  #((((     ((((((((/   (##                                 ")
+    print ("          ####  #(((  (((( (((( ((((((                                         ")
+    print ("           ##(( ((((   ((////// (((                                            ")
+    print ("            ((((((((/   ///////                                                ")
+    print ("             ((((((//    ///                                                   ")
+    print ("              ///////                                          ")
+    print ("               ////                                     ")
 
-print ("\n")
-print (bcolors.OKBLUE + "                                            @@@@@@@@   @@@@@@@@@         ")
-print ("                             @@@@@@@      @@@@@@@@@@@ @@@@@@@@@@@@             ")
-print ("                           @@@@@@@@@@@@  @@@@@@ @@@@@ @@@@@@@@@@@@@  @@@@      ")
-print ("  @@@@          @@@@@@@@@@ @@@@@@@@@@@@@ @@@@@@ @@@@  @@@@@%   @@@ @@@@@@@@@   ")
-print ("@@@@@@@@@    @@@@@@@@@@@@@ #@@@@@@@@@@@@ @@@@@@      @@@@@@@     @@@@@@@@@@@@@ ")
-print ("@@@@@@@@@@   @@@@@@@@@@@@  @@@@@  @@@@@ #&&&&&&&&    @@@@@@@@@@ @@@@@@ @@@@@@@@")
-print (" @@@@@@@@@@   @@@@@@@@@@@@  @@@@@  @@@@@ #&&&&&&&&   @@@@@@@@@@ @@@@@   @@@@@@@")
-print ("  @@@@@@@@@@&  @@@@@@@@@    @@@@&  &&&&&  &&&&&&&&% &&&&&&&&@  &@@@@    @@@@@@@")
-print ("   @@@@  %@&&&  @@@&&&       &&&&&  &&&&    &&&&&&& &&&&      &&&&@     (@@@@# ")
-print ("    @@@&   &&&&  &&&&&.      &&&&&  &&&&      %%%% %&&&      &&&&              ")
-print ("     &&&&   &&&%  &&%%%%%%%%  %%%%  %%%% /%%   %%% %%%%%%%  %&&& &&&&&&        ")
-print ("      &&&&   %%%#  %%%%%%%%#  %%%%  %%%% %%%%%%%%% %%%%%%% %%%%&&&&&&          ")
-print ("       %%%%   %%%   %######   ##### #### ########  ##%%%%  %%%%%%%&            ")
-print ("        %%%%   ###  #####      #### #### (######             %%%%              ")
-print ("         %###  ####  #((((     ((((((((/   (##                                 ")
-print ("          ####  #(((  (((( (((( ((((((                                         ")
-print ("           ##(( ((((   ((////// (((                                            ")
-print ("            ((((((((/   ///////                                                ")
-print ("             ((((((//    ///                                                   ")
-print ("              ///////                                          ")
-print ("               ////                                     ")
+    print ("\n")
+    time.sleep (0.5)
+    print (bcolors.RED + "                   +++++++++++++++++++++++++++++++++++++++                    " + bcolors.ENDC)
+    print ("                     Developed by: Luna (" + bcolors.OKGREEN + "nao3301 " + bcolors.ENDC + "&" + bcolors.OKGREEN + " lupa" + bcolors.ENDC + ")")
+    print ("                     Version: 0.2.X2 beta    ")
+    print ("                     Time: "+ bcolors.ORANGE + time_now + "  " + bcolors.ENDC)
+    print ("                     Modules: " + bcolors.MAGENTA +  str(modules) + bcolors.ENDC)
+    print (bcolors.RED + "                   +++++++++++++++++++++++++++++++++++++++                    \n" + bcolors.ENDC)
+    print ("\n\n\n")
 
-print ("\n")
-time.sleep (0.5)
-print (bcolors.RED + "                   +++++++++++++++++++++++++++++++++++++++                    " + bcolors.ENDC)
-print ("                     Developed by: Luna (" + bcolors.OKGREEN + "nao3301 " + bcolors.ENDC + "&" + bcolors.OKGREEN + " lupa" + bcolors.ENDC + ")")
-print ("                     Version: 0.2.X2 beta    ")
-print ("                     Time: "+ bcolors.ORANGE + time_now + "  " + bcolors.ENDC)
-print ("                     Modules: " + bcolors.MAGENTA +  str(modules) + bcolors.ENDC)
-print (bcolors.RED + "                   +++++++++++++++++++++++++++++++++++++++                    \n" + bcolors.ENDC)
-print ("\n\n\n")
+    time.sleep (0.5)
 
-time.sleep (0.5)
+logo()
